requirements.txt
django >= 4.1
djangorestframework
pyyaml
requests
django-cors-headers
certifi>=2021.10.8
charset-normalizer>=2.0.12
click>=8.0.4
colorama>=0.4.4
gunicorn>=20.1.0
idna>=3.3
itsdangerous>=2.1.1
